requirements.txt
Flask>=2.0
pymysql>=1.1.0
bcrypt>=4.0
cloud-sql-python-connector[pymysql]>=1.18.0
python-dotenv>=1.0.0
gunicorn>=21.0